extract inner-type
